% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{defaults}
\alias{defaults}
\alias{defaultCommand}
\alias{defaultDownloadVersion}
\alias{defaultMinimumVersion}
\alias{defaultCacheDirectory}
\title{Defaults for \pkg{biocconda}}
\usage{
defaultCommand()

defaultDownloadVersion()

defaultMinimumVersion()

defaultCacheDirectory()
}
\value{
For \code{defaultCommand}, a string specifying the expected command-line invocation of an existing conda installation.

For \code{defaultDownloadVersion}, a string specifying the version of conda to download if no existing installation can be found.

For \code{defaultMinimumVersion}, a string specifying the minimum version of an existing conda installation.

For \code{defaultCacheDirectory}, a string containing the path to the cache directory for \pkg{biocconda}-managed conda installations.
}
\description{
Defaults for \pkg{biocconda}
}
\details{
The \code{BIOCCONDA_CONDA_COMMAND} environment variable will override the default setting of \code{defaultCommand}.

The \code{BIOCCONDA_CONDA_DOWNLOAD_VERSION} environment variable will override the default setting of \code{defaultDownloadVersion}.

The \code{BIOCCONDA_CONDA_MINIMUM_VERSION} environment variable will override the default setting of \code{defaultMinimumVersion}.

The \code{BIOCCONDA_CONDA_CACHE_DIRECTORY} environment variable will override the default setting of \code{defaultCacheDirectory}.
}
\examples{
defaultCommand()
defaultDownloadVersion()
defaultMinimumVersion()
defaultCacheDirectory()

}
\author{
Aaron Lun
}
