% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useSystemDir.R
\name{destroyOldVersions}
\alias{destroyOldVersions}
\title{Destroy old versions?}
\usage{
destroyOldVersions()
}
\value{
Logical scalar providing an answer to the above.
}
\description{
Should we destroy old installations of Conda from previous versions of \pkg{basilisk} 
(or old environment installations, for \pkg{basilisk} client packages)?
}
\details{
The default value is \code{TRUE}, in order to save some hard drive space.
This can be changed by setting \code{BASILISK_NO_DESTROY} environment variable to \code{"1"}.
}
\seealso{
\code{\link{installConda}}, where this function is used.

\code{\link{clearObsoleteDir}}, which may be triggered by this function.
}
\author{
Aaron Lun
}
