% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBasiliskDir.R
\name{getBasiliskDir}
\alias{getBasiliskDir}
\title{Get the \pkg{basilisk} Anaconda directory}
\usage{
getBasiliskDir(assume.installed = TRUE)
}
\arguments{
\item{assume.installed}{Logical scalar indicating whether we can assume the Anaconda instance is already installed.}
}
\value{
String containing the path to the Anaconda instance.
If \code{assume.installed=TRUE}, this function will throw an error if the expected directory does not exist.
}
\description{
Find the installation directory for the \pkg{basilisk}-managed Anaconda instance.
}
\details{
By default, Anaconda is installed to a location specified by \code{\link{getExternalDir}}.
This ensures that R package build systems do not attempt to generate binaries that include the Anaconda installation;
such binaries are not relocatable due to the presence of hard-coded paths, resulting in run-time failures.

If the \code{BASILISK_EXTERNAL_ANACONDA} environment variable is set to a path to an existing Anaconda installation,
the function will return it directly without modification.
This allows users to use their own Anaconda instances with \pkg{basilisk} but, 
in turn, they are responsible for managing it.

If the \code{BASILISK_USE_SYSTEM_DIR} environment variable is set to \code{"1"},
the function will return a path to a location inside the \code{basilisk} system installation directory.
This is the ideal approach when installing from source as any Anaconda and \pkg{basilisk} re-installations are synchronized.
It also ensures that any R process that can load \pkg{basilisk} will also have permissions to access the Anaconda instance,
which makes life easier for sysadmins of clusters or other shared resources.
}
\examples{
# Setting the environment variable to run this example: 
# all other modes rely on installation of basilisk.
old <- Sys.getenv("BASILISK_USE_SYSTEM_DIR")
Sys.setenv(BASILISK_USE_SYSTEM_DIR=1)

getBasiliskDir(assume.installed=FALSE)

Sys.setenv(BASILISK_USE_SYSTEM_DIR=old)
}
\author{
Aaron Lun
}
