% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getBinaries}
\alias{getBinaries}
\alias{getCondaBinary}
\alias{getPythonBinary}
\title{Get binary paths}
\usage{
getCondaBinary(loc)

getPythonBinary(loc)
}
\arguments{
\item{loc}{String containing the path to the root of a conda instance or environment.}
}
\value{
String containing the path to the conda or Python executable inside \code{loc}.
If \code{loc} is not supplied, the relative path from the root of the environment is returned.
}
\description{
Get binary paths
}
\details{
This code is largely copied from \pkg{reticulate},
and is only present here as they do not export these utilities for general consumption.
}
\examples{
getCondaBinary()

getPythonBinary()
}
\author{
Aaron Lun
}
