% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExternalDir.R
\name{getExternalDir}
\alias{getExternalDir}
\title{Get external Anaconda installation}
\usage{
getExternalDir()
}
\value{
String containing a path to an appropriate external folder.
}
\description{
Define an external location for installing the Anaconda instance, 
usually on Windows or MacOSX.
}
\details{
We add the version information so that re-installation of \pkg{basilisk} will install a new instance of Anaconda.
(This assumes that \pkg{basilisk} and \pkg{basilisk.utils} have synchronized version bumps.)
See \code{\link{installAnaconda}} for more details on how this external location is managed.

This function will respect any setting of \code{BASILISK_EXTERNAL_DIR}, 
see \code{?\link{installAnaconda}} for more details.
}
\examples{
getExternalDir()

}
\seealso{
\code{\link{basiliskAnaconda}}, where this function is used for Windows or MacOSX.
}
\author{
Aaron Lun
}
