% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getLockFile}
\alias{getLockFile}
\title{Get lock file}
\usage{
getLockFile(path)
}
\arguments{
\item{path}{String containing a path to an Anaconda installation or conda environment that is to be installed.}
}
\value{
String containing a path to a lock file,
to be \code{touch}ed before installation starts and deleted after installation finishes \emph{successfully}.
}
\description{
Get the path to a lock file, typically used to mark an installation in progress 
(or an incomplete installation that needs to be replaced).
}
\examples{
getLockFile("AAAA")

}
\author{
Aaron Lun
}
