% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSystemDir.R
\name{getSystemDir}
\alias{getSystemDir}
\title{Get the system installation directory}
\usage{
getSystemDir(pkgname, installed)
}
\arguments{
\item{pkgname}{String containing the package name.}

\item{installed}{Logical scalar specifying whether the package is likely to be installed yet.}
}
\value{
String containing the path to the (likely, if \code{installed=FALSE}) installation directory for \code{pkgname}.
}
\description{
Get the system installation directory for a package.
This is not entirely trivial as it may be called before the package is installed.
}
\examples{
getSystemDir("basilisk", installed=FALSE)

}
\author{
Aaron Lun
}
