% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useSystemDir.R
\name{noFallbackR}
\alias{noFallbackR}
\title{Don't install a fallback version of R}
\usage{
noFallbackR()
}
\value{
Logical scalar providing an answer to the above.
}
\description{
By default, \code{\link{installConda}} will install a conda-managed R in a new environment,
to provide a fallback for \code{GLIBCXX}-mismatch errors in C++-dependent Python packages like \pkg{scipy}.
This cost can be avoided if the version of the C++ standard library used by R is known to be compatible with client environments.
}
\details{
The default value is \code{FALSE} in order to provide a fallback when clients observe \code{GLIBCXX} errors.
This can be changed by setting the \code{BASILISK_NO_FALLBACK_R} environment variable to \code{"1"}.
}
\seealso{
\code{\link{installConda}}, where this function is used.
}
\author{
Aaron Lun
}
