% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setVariable.R
\name{setVariable}
\alias{setVariable}
\title{Set an environment variable}
\usage{
setVariable(name, value)
}
\arguments{
\item{name}{String containing the name of an environment variable.}

\item{value}{String containing the value of an environment variable.
This can be \code{NA} to unset the variable.}
}
\value{
String containing the value of the variable before running this function;
or \code{NA}, if the variable was not set.
}
\description{
Set an environment variable safely, unsetting it if the supplied value is \code{NA}.
}
\author{
Aaron Lun
}
