% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useSystemDir.R
\name{useMambaForge}
\alias{useMambaForge}
\title{Use Mamba-forge instead of Miniconda?}
\usage{
useMambaForge()
}
\value{
Logical scalar providing an answer to the above.
}
\description{
Should we use Mamba-forge instead of Miniconda?
The former is fully open-source and is more amenable for commercial environments.
}
\details{
The default value is \code{FALSE} for backwards compatibility.
This can be changed by setting \code{BASILISK_USE_MAMBAFORGE} environment variable to \code{"1"}.
}
\seealso{
\url{https://github.com/conda-forge/miniforge}, for more details.
}
\author{
Aaron Lun
}
