% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{useSystemDir}
\alias{useSystemDir}
\title{Use the R system directory?}
\usage{
useSystemDir()
}
\value{
Logical scalar providing an answer to the above.
}
\description{
Should we use the R system directory for installing \pkg{basilisk}'s Anaconda instance or client environments?
}
\seealso{
\code{\link{getBasiliskDir}} and \code{\link{getEnvironmentDir}}, where these functions are used.
}
\author{
Aaron Lun
}
