% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clearExternalDir.R
\name{clearExternalDir}
\alias{clearExternalDir}
\title{Clear the external installation directory}
\usage{
clearExternalDir()
}
\value{
All Anaconda instances and environments of the same Bioconductor release 
as the current \pkg{basilisk} installation are destroyed.
}
\description{
Clear the external installation directory by removing all Anaconda instances 
installed for different versions of \pkg{basilisk} with the same middle version number
(i.e., same Bioconductor release).
}
\examples{
# We can't actually run clearExternalDir() here, as it 
# relies on basilisk already being installed.
print("dummy test to pass BiocCheck")

}
\seealso{
\code{\link{getExternalDir}}, which determines the location of the external directory.

\code{\link{installAnaconda}}, for the motivation behind this function.
}
\author{
Aaron Lun
}
