% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBasiliskDir.R
\name{getBasiliskDir}
\alias{getBasiliskDir}
\title{Get the \pkg{basilisk} Anaconda directory}
\usage{
getBasiliskDir(mustWork = TRUE)
}
\arguments{
\item{mustWork}{Logical scalar indicating whether to throw an error if the directory does not exist.}
}
\value{
String containing the path to the Anaconda instance.
}
\description{
Find the installation directory for the \pkg{basilisk}-managed Anaconda instance.
}
\details{
If the \code{BASILISK_USE_ANACONDA} environment variable is set,
the function will return it directly without modification.
This allows users to use their own Anaconda instances with \pkg{basilisk} but, 
in turn, they are responsible for managing it.
}
\author{
Aaron Lun
}
