% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBasiliskDir.R
\name{getBasiliskDir}
\alias{getBasiliskDir}
\title{Get the \pkg{basilisk} Anaconda directory}
\usage{
getBasiliskDir(assume.installed = TRUE)
}
\arguments{
\item{assume.installed}{Logical scalar indicating whether we can assume the Anaconda instance is already installed.}
}
\value{
String containing the path to the Anaconda instance.
}
\description{
Find the installation directory for the \pkg{basilisk}-managed Anaconda instance.
}
\details{
If the \code{BASILISK_USE_ANACONDA} environment variable is set,
the function will return it directly without modification.
This allows users to use their own Anaconda instances with \pkg{basilisk} but, 
in turn, they are responsible for managing it.

If \code{assume.installed=TRUE}, this function will throw an error if the Anaconda directory does not exist.
On Unix, it will also use \code{\link{system.file}} to ensure that the correct directory is recovered,
even if \code{\link{.libPaths}} is altered to some other installation location.
}
\examples{
getBasiliskDir(assume.installed=FALSE)

}
\author{
Aaron Lun
}
