% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExternalDir.R
\name{getExternalDir}
\alias{getExternalDir}
\title{Get external Miniconda installation}
\usage{
getExternalDir()
}
\value{
String containing a path to an appropriate external folder.
}
\description{
Define an external location for installing the Miniconda instance and \pkg{basilisk} environments.
}
\details{
We add the version information so that re-installation of \pkg{basilisk} will install a new instance of Miniconda.
(This assumes that \pkg{basilisk} and \pkg{basilisk.utils} have synchronized version bumps.)
See \code{\link{installMiniconda}} for more details on how this external location is managed.

If the \code{BASILISK_EXTERNAL_DIR} environment variable is set to some location,
this will be used instead as the installation directory.
Setting this variable is occasionally necessary if the default path returned by \code{\link{user_cache_dir}} has spaces;
or on Windows, if the 260 character limit is exceeded after combining the default path with deeply nested Miniconda paths.
}
\examples{
# We can't actually run getExternalDir() here, as it 
# either relies on basilisk already being installed.
print("dummy test to pass BiocCheck")

}
\seealso{
\code{\link{getBasiliskDir}}, where this function is used for Windows or MacOSX.
}
\author{
Aaron Lun
}
