% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installAnaconda.R
\name{installAnaconda}
\alias{installAnaconda}
\title{Install Anaconda}
\usage{
installAnaconda()
}
\value{
An Anaconda instance is created at the location specified by \code{\link{getBasiliskDir}}.
}
\description{
Install Anaconda (version 3, 2019.10) to a destination path that depends on the operating system.
This skips the installation if said path already exists.
}
\details{
This function was originally created from code in \url{https://github.com/hafen/rminiconda},
also borrowing code from \pkg{reticulate}'s \code{install_miniconda} for correct Windows installation.
We use \pkg{BiocFileCache} if available to avoid redownloading the Anaconda installer upon \pkg{basilisk} re-installation.

Whenever \code{installAnaconda} is re-run (and \code{BASILISK_USE_SYSTEM_DIR} is not set, see \code{?\link{getBasiliskDir}}),
the previous Anaconda installation and its various \pkg{basilisk} environments are destroyed.
This avoids duplication of Anaconda instances that would otherwise chew up disk space at 3 GB a pop.

After the destruction of the previous instance, we rely on the client packages to recreate their required environments.
They should do this automatically if they are using \pkg{basilisk} correctly.

Users can disable this destruction by setting the \code{BASILISK_NO_DESTROY} environment variable to \code{"1"}.
This may be necessary on rare occasions when running multiple R instances on the same Bioconductor release.
(Setting this variable is not required for instances using different Bioconductor releases.)
}
\examples{
# We can't actually run installAnaconda() here, as it 
# either relies on basilisk already being installed or
# it has a hard-coded path to the basilisk system dir.
print("dummy test to pass BiocCheck")

}
\author{
Aaron Lun
}
