% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isWindows.R
\name{isWindows}
\alias{isWindows}
\alias{isMacOSX}
\alias{isMacOSXArm}
\alias{isLinux}
\alias{isLinuxAarch64}
\title{Find the operating system or architecture.}
\usage{
isWindows()

isMacOSX()

isMacOSXArm()

isLinux()

isLinuxAarch64()
}
\value{
Logical scalar indicating whether we are on the specified OS and/or architecture.
}
\description{
Indicate whether we are on Windows or MacOSX.
For MacOSX, we can also determine if we are on an x86-64 or Arm-based architecture.
}
\examples{
isWindows()
isMacOSX()
isLinux()
}
\author{
Aaron Lun
}
