% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isWindows.R
\name{isWindows}
\alias{isWindows}
\alias{isMacOSX}
\title{Find the operating system}
\usage{
isWindows()

isMacOSX()
}
\value{
Logical scalar indicating whether we are on the specified OS.
}
\description{
Indicate whether we are on Windows or MacOSX.
}
\examples{
isWindows()
isMacOSX()
}
\author{
Aaron Lun
}
