% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useSystemDir.R
\name{useSystemDir}
\alias{useSystemDir}
\title{Use the R system directory?}
\usage{
useSystemDir()
}
\value{
Logical scalar providing an answer to the above.
}
\description{
Should we use the R system directory for installing \pkg{basilisk}'s Conda instance (or client environments)?
}
\details{
The default value is \code{FALSE} to avoid problems with position-dependent code in packaged binaries.
This can be changed by setting \code{BASILISK_USE_SYSTEM_DIR} environment variable to \code{"1"}.
}
\seealso{
\code{\link{installConda}} and \code{\link{getCondaDir}}, where this function is used.
}
\author{
Aaron Lun
}
